fix tests

+++ rust/leap/src/lib.rs
@@ -1,3 +1,3 @@
-pub fn is_leap_year(year: u16) -> bool {
+pub fn is_leap_year(year: u64) -> bool {
     year % 4 == 0 && (year % 100 != 0 || year % 400 == 0)
 }
